update by josephine morre
first update of epen
copy by chanos
